minor fix + refactor: Fix user controller import and test cases
