+++ TESTING.md
@@ -1,0 +1,44 @@
+# Testing Guide
+
+### Isolate the Unit:
+
+A unit test should focus on a single, independent unit of code (e.g., a function, a class method). Use mocks, stubs, or fakes to simulate dependencies and ensure the unit's behavior is tested in isolation, without relying on external systems like databases or APIs.
+
+
+### Fast Execution:
+Unit tests should run quickly to provide rapid feedback during development. Avoid slow operations or external calls that can hinder test suite performance.
+
+
+### Self-Validating:
+Tests should clearly indicate success or failure without manual intervention. The test output should be unambiguous.
+
+### Repeatable:
+Tests should produce the same results every time they are run, regardless of the environment or execution order.
+
+### Timely:
+Write unit tests frequently and early in the development process, ideally alongside the code being tested (Test-Driven Development).
+
+
+### Understandable:
+Test code should be clear, concise, and easy to understand, even for someone unfamiliar with the system. Use descriptive test names.
+
+### No Side Effects:
+Unit tests should not alter the state of the system or leave behind artifacts that could affect subsequent tests.
+
+
+### Rules for Writing Integration Tests:
+
+### Test Interactions:
+Integration tests verify the interactions and communication between different components or modules of a system, or between the system and external dependencies (e.g., databases, APIs, message queues).
+### Realistic Environment:
+While unit tests strive for isolation, integration tests often require a more realistic environment, including actual dependencies, to accurately simulate real-world interactions.
+### Focus on Interfaces:
+Test the interfaces and contracts between components to ensure they work together as expected.
+### Manage Dependencies:
+Carefully manage and set up the necessary infrastructure and data for integration tests. This might involve creating test databases or configuring external services.
+### Consider Performance:
+Integration tests are typically slower than unit tests due to their reliance on external systems. Optimize test setup and execution to minimize run times where possible.
+### Clear Scope:
+Define the scope of each integration test to ensure it focuses on a specific set of interactions.
+### Clean Up:
+Ensure that integration tests clean up any created data or resources after execution to maintain a consistent test environment.
